shadow in news bottom sheet
