+++ src/components/BoxShadow/index.ts
@@ -1,0 +1,2 @@
+export * from "./BoxShadowView"
+export * from "./BoxShadowCanvas"
